Convert reboot resource to a custom resource with descriptions
This is a simple resource. It's way easier to understand in a single file.
