+++ lib/chef/providers.rb
@@ -43,7 +43,6 @@ require_relative "provider/noop"
 require_relative "provider/package"
 require_relative "provider/powershell_script"
 require_relative "provider/osx_profile"
-require_relative "provider/reboot"
 require_relative "provider/remote_directory"
 require_relative "provider/remote_file"
 require_relative "provider/route"
